+++ tests/testthat/test_parse_tex.R
@@ -41,5 +41,13 @@ test_that("Parse tiny documents", {
                    
 })
 
+test_that("parse2", {
+  library(data.table)
+  y <- parse_tex("SchoolFunding/SchoolFunding.tex")
+  y2 <- parse_tex2("SchoolFunding/SchoolFunding.tex")
+  expect_equal(y, y2, check.attributes = FALSE)
+})
+
+